chore: refactor dashboard server
