+++ dashboard/server/api/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+    testSocketConnection: (client) => {
+        client.on('test-conn', (cb) => {
+            cb('hello world');
+        });
+    },
+};
